refactor: load image 메서드 추상화
